Actualización Graphs

+++ DocCare_Tracker/app/src/main/java/com/example/doccare_tracker/model/Informacion_Personal/Usuario/ModificarDatosUsuario.kt
@@ -1,13 +1,13 @@
 package com.example.doccare_tracker.model.Informacion_Personal.Usuario
 
 data class ModificarDatosUsuario(
-    val altura: Int,
+    val altura: Float,
     val apellidom: String,
     val apellidop: String,
-    val cintura: Int,
+    val cintura: Float,
     val clave: String,
     val correo: String,
     val nombre: String,
-    val peso: Int,
+    val peso: Float,
     val usuario_id: Int
 )
