+++ DocCare_Tracker/app/src/main/java/com/example/doccare_tracker/model/Registros/Usuario/RegistrarDetalleUsuario.kt
@@ -1,10 +1,10 @@
 package com.example.doccare_tracker.model.Registros.Usuario
 
 data class RegistrarDetalleUsuario(
-    val altura: Int,
-    val circunferencia_abdominal: Int,
+    val altura: Float,
+    val circunferencia_abdominal: Float,
     val clave_unica: String,
-    val peso: Int,
+    val peso: Float,
     val sexo: String,
     val usuario_id: Int
 )
